+++ src/commonMain/kotlin/org/hisp/dhis/rules/models/RuleEngineValidationException.kt
@@ -1,3 +1,3 @@
-package org.hisp.dhis.rules.engine
+package org.hisp.dhis.rules.models
 
 class RuleEngineValidationException(cause: IllegalArgumentException) : IllegalArgumentException(cause.message)
